// src/auth/dto/google-account-response.dto.ts
export interface GoogleAccountResponse {
  iss: string;
  azp: string;
  aud: string;
  sub: string;
  email: string;
  email_verified: string;
  name: string;
  picture: string;
  given_name: string;
  family_name: string;
  iat: string;
  exp: string;
  alg: string;
  kid: string;
  typ: string;
}